backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
